+++ src/main/java/edu/ie3/datamodel/io/source/IdCoordinateSource.java
@@ -13,7 +13,9 @@ import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.commons.lang3.tuple.Pair;
+import javax.measure.quantity.Length;
 import org.locationtech.jts.geom.Point;
+import tech.units.indriya.ComparableQuantity;
 
 /**
  * This class serves mapping purposes between the ID of a coordinate and the actual coordinate with
@@ -87,12 +89,25 @@ public class IdCoordinateSource {
   }
 
   /**
-   * Returns the nearest n coordinate points to the given coordinate from a collection of all
-   * available points
+   * Returns the nearest n coordinate points. If n is greater than four, this method will try to
+   * return the corner points of the bounding box.
+   *
+   * @param coordinate the coordinate to look up
+   * @param n number of searched points
+   * @return the nearest n coordinates or all coordinates if n is less than all available points
+   */
+  List<CoordinateDistance> getNearestCoordinates(Point coordinate, int n);
+
+  /**
+   * Returns the closest n coordinate points to the given coordinate, that are inside a given
+   * bounding box, from a collection of all available points. The bounding box is calculated with
+   * the given distance. If n is greater than four, this method will try to return the corner points
+   * of the bounding box.
    *
    * @param coordinate the coordinate to look up the nearest neighbours for
    * @param n how many neighbours to look up
-   * @return the n nearest coordinates to the given point
+   * @param distance to the borders of the envelope that contains the coordinates
+   * @return the nearest n coordinates to the given point
    */
   public List<CoordinateDistance> getNearestCoordinates(Point coordinate, int n) {
     return getNearestCoordinates(coordinate, n, getAllCoordinates());
@@ -127,21 +142,94 @@ public class IdCoordinateSource {
   }
 
   /**
-   * Returns the nearest n coordinate points to the given coordinate from a given collection of
-   * points. If the set is empty or null we look through all coordinates.
+   * Calculates and returns the nearest n coordinate distances to the given coordinate from a given
+   * collection of points. If the set is empty or null an empty list is returned. If n is greater
+   * than four, this method will try to return the corner points of the bounding box.
    *
    * @param coordinate the coordinate to look up the nearest neighbours for
    * @param n how many neighbours to look up
    * @param coordinates the collection of points
-   * @return the n nearest coordinates to the given point
+   * @return a list of the nearest n coordinates to the given point or an empty list
    */
   public List<CoordinateDistance> getNearestCoordinates(
       Point coordinate, int n, Collection<Point> coordinates) {
-    SortedSet<CoordinateDistance> sortedDistances =
-        GeoUtils.calcOrderedCoordinateDistances(
-            coordinate,
-            (coordinates != null && !coordinates.isEmpty()) ? coordinates : getAllCoordinates());
-    return sortedDistances.stream().limit(n).toList();
+    if (coordinates != null && !coordinates.isEmpty()) {
+      SortedSet<CoordinateDistance> sortedDistances =
+          GeoUtils.calcOrderedCoordinateDistances(coordinate, coordinates);
+      return restrictToBoundingBox(coordinate, sortedDistances, n);
+    } else {
+      return Collections.emptyList();
+    }
+  }
+
+  /**
+   * Method for evaluating the found points. This method tries to return the four corner points of
+   * the bounding box of the given coordinate. If one of the found points matches the given
+   * coordinate, only this point is returned. If the given number of searched points is less than
+   * four, this method will only return the nearest n corner points. If the given number of searched
+   * points is greater than four, this method will return the four corner points plus the nearest n
+   * points to match the number of searched points.
+   *
+   * <p>To work properly, the given collection of {@link CoordinateDistance}'s should already be
+   * sorted by distance.
+   *
+   * @param coordinate at the center of the bounding box
+   * @param distances list of found points with their distances
+   * @param numberOfPoints that should be returned
+   * @return list of distances
+   */
+  default List<CoordinateDistance> restrictToBoundingBox(
+      Point coordinate, Collection<CoordinateDistance> distances, int numberOfPoints) {
+    boolean topLeft = false;
+    boolean topRight = false;
+    boolean bottomLeft = false;
+    boolean bottomRight = false;
+
+    List<CoordinateDistance> resultingDistances = new ArrayList<>();
+    List<CoordinateDistance> other = new ArrayList<>();
+
+    // search for smallest bounding box
+    for (CoordinateDistance distance : distances) {
+      Point point = distance.getCoordinateB();
+
+      // check for bounding box
+      if (!topLeft && (point.getX() < coordinate.getX() && point.getY() > coordinate.getY())) {
+        resultingDistances.add(distance);
+        topLeft = true;
+      } else if (!topRight
+          && (point.getX() > coordinate.getX() && point.getY() > coordinate.getY())) {
+        resultingDistances.add(distance);
+        topRight = true;
+      } else if (!bottomLeft
+          && (point.getX() < coordinate.getX() && point.getY() < coordinate.getY())) {
+        resultingDistances.add(distance);
+        bottomLeft = true;
+      } else if (!bottomRight
+          && (point.getX() > coordinate.getX() && point.getY() < coordinate.getY())) {
+        resultingDistances.add(distance);
+        bottomRight = true;
+      } else if (coordinate.equalsExact(point, 1e-6)) {
+        // if current point is matching the given coordinate, we need to return only the current
+        // point
+
+        resultingDistances.clear();
+        resultingDistances.add(distance);
+        return resultingDistances;
+      } else {
+        other.add(distance);
+      }
+    }
+
+    // check if n distances are found
+    int diff = numberOfPoints - resultingDistances.size();
+
+    if (diff > 0) {
+      resultingDistances.addAll(other.stream().limit(diff).toList());
+    } else if (diff < 0) {
+      return resultingDistances.stream().limit(numberOfPoints).toList();
+    }
+
+    return resultingDistances;
   }
 
   public int getCoordinateCount() {