+++ src/main/java/edu/ie3/datamodel/io/source/couchbase/CouchbaseWeatherSource.java
@@ -44,35 +44,6 @@ public class CouchbaseWeatherSource extends WeatherSource {
   /**
    * Instantiate a weather source utilising a connection to a couchbase instance obtained via the
    * connector. This convenient constructor uses {@link CouchbaseWeatherSource#DEFAULT_KEY_PREFIX}
-   * as key prefix and {@link CouchbaseWeatherSource#DEFAULT_TIMESTAMP_PATTERN} as timestamp
-   * pattern.
-   *
-   * @param connector Connector, that establishes the connection to the couchbase instance
-   * @param coordinateSource Source to obtain actual coordinates from
-   * @param coordinateIdColumnName Name of the column containing the information about the
-   *     coordinate identifier
-   * @param weatherFactory Factory to transfer field to value mapping into actual java object
-   *     instances
-   * @deprecated Use {@link CouchbaseWeatherSource#CouchbaseWeatherSource(CouchbaseConnector,
-   *     IdCoordinateSource, String, TimeBasedWeatherValueFactory, String)} instead
-   */
-  @Deprecated(since = "3.0", forRemoval = true)
-  public CouchbaseWeatherSource(
-      CouchbaseConnector connector,
-      IdCoordinateSource coordinateSource,
-      String coordinateIdColumnName,
-      TimeBasedWeatherValueFactory weatherFactory) {
-    this(
-        connector,
-        coordinateSource,
-        coordinateIdColumnName,
-        weatherFactory,
-        DEFAULT_TIMESTAMP_PATTERN);
-  }
-
-  /**
-   * Instantiate a weather source utilising a connection to a couchbase instance obtained via the
-   * connector. This convenient constructor uses {@link CouchbaseWeatherSource#DEFAULT_KEY_PREFIX}
    * as key prefix.
    *
    * @param connector Connector, that establishes the connection to the couchbase instance
@@ -96,35 +67,6 @@ public class CouchbaseWeatherSource extends WeatherSource {
         DEFAULT_KEY_PREFIX,
         weatherFactory,
         timeStampPattern);
-  }
-
-  /**
-   * Instantiate a weather source utilising a connection to a couchbase instance obtained via the
-   * connector. This convenient constructor uses {@link
-   * CouchbaseWeatherSource#DEFAULT_TIMESTAMP_PATTERN} as timestamp pattern.
-   *
-   * @param connector Connector, that establishes the connection to the couchbase instance
-   * @param coordinateSource Source to obtain actual coordinates from
-   * @param keyPrefix Prefix of entries, that belong to weather
-   * @param weatherFactory Factory to transfer field to value mapping into actual java object
-   *     instances
-   * @deprecated Use {@link CouchbaseWeatherSource#CouchbaseWeatherSource(CouchbaseConnector,
-   *     IdCoordinateSource, String, String, TimeBasedWeatherValueFactory, String)} instead
-   */
-  @Deprecated(since = "3.0", forRemoval = true)
-  public CouchbaseWeatherSource(
-      CouchbaseConnector connector,
-      IdCoordinateSource coordinateSource,
-      String coordinateIdColumnName,
-      String keyPrefix,
-      TimeBasedWeatherValueFactory weatherFactory) {
-    this(
-        connector,
-        coordinateSource,
-        coordinateIdColumnName,
-        keyPrefix,
-        weatherFactory,
-        DEFAULT_TIMESTAMP_PATTERN);
   }
 
   /**