+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvJointGridContainerSource.java
@@ -5,7 +5,10 @@
 */
 package edu.ie3.datamodel.io.source.csv;
 
+import edu.ie3.datamodel.exceptions.FileException;
 import edu.ie3.datamodel.exceptions.SourceException;
+import edu.ie3.datamodel.io.naming.DefaultDirectoryHierarchy;
+import edu.ie3.datamodel.io.naming.EntityPersistenceNamingStrategy;
 import edu.ie3.datamodel.io.naming.FileNamingStrategy;
 import edu.ie3.datamodel.io.source.*;
 import edu.ie3.datamodel.models.input.container.GraphicElements;
@@ -17,12 +20,23 @@ import edu.ie3.datamodel.models.input.container.SystemParticipants;
 public class CsvJointGridContainerSource {
   private CsvJointGridContainerSource() {}
 
-  public static JointGridContainer read(String gridName, String csvSep, String directoryPath)
-      throws SourceException {
+  public static JointGridContainer read(
+      String gridName, String csvSep, String directoryPath, boolean isHierarchic)
+      throws SourceException, FileException {
 
     /* Parameterization */
+    FileNamingStrategy namingStrategy;
 
-    FileNamingStrategy namingStrategy = new FileNamingStrategy(); // Default naming strategy
+    if (isHierarchic) {
+      // Hierarchic structure
+      DefaultDirectoryHierarchy fileHierarchy =
+          new DefaultDirectoryHierarchy(directoryPath, gridName);
+      namingStrategy = new FileNamingStrategy(new EntityPersistenceNamingStrategy(), fileHierarchy);
+      fileHierarchy.validate();
+    } else {
+      // Flat structure
+      namingStrategy = new FileNamingStrategy();
+    }
 
     CsvDataSource dataSource = new CsvDataSource(csvSep, directoryPath, namingStrategy);
 
@@ -45,7 +59,7 @@ public class CsvJointGridContainerSource {
             .orElseThrow(
                 () -> new SourceException("Error during reading of system participant data."));
     GraphicElements graphicElements =
-        graphicsSource
+        graphicSource
             .getGraphicElements()
             .orElseThrow(() -> new SourceException("Error during reading of graphic elements."));